requirements.txt
streamlit==1.32.0
google-generativeai==0.3.2
python-dotenv==1.0.1
faiss-cpu==1.7.4
numpy==1.26.4
pandas==2.2.1
playwright==1.42.0
python-docx==1.1.0
PyPDF2==3.0.1
SQLAlchemy==2.0.28
bcrypt==4.1.2
python-jose==3.3.0
fastapi==0.110.0
uvicorn==0.27.1
pytest==8.0.2
black==24.2.0
flake8==7.0.0 